Move LoginView back to webgateway/views.py
